fix copyright spec compliance
A and B already means that the file contains code under A and code under B
'A and B' is not a distinct license it is rather a composite of
License: A
and
License: B

defining A and B therefore creats a duplicated definition if both are also
defined standalone.

change the copyright accordingly. the combo license stanzas are no more
and the previous or statement is now an and statement as the files in
question are not dual licensed in an either or fashion but parts are simply
one and other parts the other