do not force rebuild the tree
this blows up build time but in fact adt-run is meant to handle this.
adt-run would build the tree if it is unbuilt, an already built tree is
simply not usable with adt-run and cmake as cmake builds are not
relocatable.
iff we were to want to relocate them (which I think we do not) the way to
go about it is to figure out if the tree was relocated and only then
force rebuild it. TBH though I think this is useless work, when passing
an already built tree into adt-run there really should be no assumption
of it working. conversely, during development you do not need to use
adt-run but could just run the testsuite script manually